[update]: build error 해결 시도2
